Revert "split query proxy methods into their own spec, add delete_all for qp"
