+++ lib/neo4j/active_node/query/query_proxy.rb
@@ -28,10 +28,6 @@ module Neo4j
           @node_var || :result
         end
 
-        def rel_identity
-          @rel_var
-        end
-
         def enumerable_query(node, rel = nil)
           pluck_this = rel.nil? ? [node] : [node, rel]
           return self.pluck(*pluck_this) if @association.nil? || caller.nil?
